Update what-is-cloud-sync.md
Based on information in https://learn.microsoft.com/en-us/entra/identity/hybrid/cloud-sync/how-to-configure-entra-to-active-directory , Group Writeback will be supported in CloudSync and hence I have added the support in the Comparision Table - Group Writeback feature for Cloud sync.

I also suggest that we add an asterisk to indicate that support for Group Writeback will be deprecated in Connect Sync 

"The public preview of Group Writeback v2 in Microsoft Entra Connect Sync will no longer be available after June 30, 2024. This feature will be discontinued on this date, and you will no longer be supported in Connect Sync to provision cloud security groups to Active Directory.

We offer similar functionality in Microsoft Entra Cloud Sync called Group Provision to Active Directory that you can use instead of Group Writeback v2 for provisioning cloud security groups to Active Directory. We're working on enhancing this functionality in Cloud Sync along with other new features that we're developing in Cloud Sync."